fix: xiaomi youpin, close #4108
